requirements/base.txt
six==1.10.0
Django==2.2.4
djangorestframework==3.10.2
django-cors-headers==2.4.0
django-filter==2.0.0
pytz==2019.2
pika==1.1.0
